requirements.txt
Flask
Flask-Cors
google-genai
Flask-CORS
google-generativeai
python-dotenv
Pillow

